Move TokenToJson to service level
Avoid lower level package dependency on external generated
REST API.

Related to fabric8-services/fabric8-wit#1395
