+++ login/service_whitebox_test.go
@@ -1,11 +1,14 @@
 package login
 
 import (
+	"context"
+	"encoding/json"
 	"testing"
 
 	"github.com/fabric8-services/fabric8-auth/account"
 	"github.com/fabric8-services/fabric8-auth/resource"
 	"github.com/fabric8-services/fabric8-auth/token"
+	"golang.org/x/oauth2"
 
 	_ "github.com/lib/pq"
 	"github.com/stretchr/testify/assert"
@@ -34,4 +37,35 @@ func TestFillUserDoesntOverwriteExistingImageURL(t *testing.T) {
 	assert.Equal(t, "new email", identity.User.Email)
 	assert.Equal(t, "new username", identity.Username)
 	assert.Equal(t, "http://vpupkin.io/image.jpg", identity.User.ImageURL)
+}
+
+func TestEncodeTokenOK(t *testing.T) {
+	accessToken := "accessToken%@!/\\&?"
+	refreshToken := "refreshToken%@!/\\&?"
+	tokenType := "tokenType%@!/\\&?"
+	expiresIn := 1800
+	var refreshExpiresIn float64
+	refreshExpiresIn = 2.59e6
+
+	outhToken := &oauth2.Token{
+		AccessToken:  accessToken,
+		RefreshToken: refreshToken,
+		TokenType:    tokenType,
+	}
+	extra := map[string]interface{}{
+		"expires_in":         expiresIn,
+		"refresh_expires_in": refreshExpiresIn,
+	}
+	tokenJson, err := TokenToJson(context.Background(), outhToken.WithExtra(extra))
+	assert.Nil(t, err)
+	b := []byte(tokenJson)
+	tokenData := &token.TokenSet{}
+	err = json.Unmarshal(b, tokenData)
+	assert.Nil(t, err)
+
+	assert.Equal(t, accessToken, *tokenData.AccessToken)
+	assert.Equal(t, refreshToken, *tokenData.RefreshToken)
+	assert.Equal(t, tokenType, *tokenData.TokenType)
+	assert.Equal(t, int64(expiresIn), *tokenData.ExpiresIn)
+	assert.Equal(t, int64(refreshExpiresIn), *tokenData.RefreshExpiresIn)
 }
